fix: Move LLMButtons component to theme
