+++ src/theme/DocItem/Content/index.js
@@ -2,11 +2,10 @@ import { useDoc } from '@docusaurus/plugin-content-docs/client';
 import { useLocation } from '@docusaurus/router';
 import { ThemeClassNames } from '@docusaurus/theme-common';
 import Heading from '@theme/Heading';
+import LLMButtons from '@theme/LLMButtons';
 import MDXContent from '@theme/MDXContent';
 import clsx from 'clsx';
 import React from 'react';
-
-import LLMButtons from '../../../components/LLMButtons';
 
 function useSyntheticTitle() {
   const { metadata, frontMatter, contentTitle } = useDoc();
